Add more tests for feature extraction, training, and classification. Reorganize vision_backend task test files.
